Update
- Bus error fix in itoa & strtrim
- Better protect for ft_*_fd functions
- Change var type from int to size_t in some function
- Good free in split

+++ ft_lstsize_bonus.c
@@ -6,7 +6,7 @@
 /*   By: siferrar <siferrar@student.le-101.fr>      +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/10/19 02:37:14 by siferrar     #+#   ##    ##    #+#       */
-/*   Updated: 2019/10/21 15:23:40 by siferrar    ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/10/28 15:17:48 by siferrar    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -15,7 +15,7 @@
 
 int	ft_lstsize(t_list *lst)
 {
-	int lenght;
+	size_t lenght;
 
 	lenght = 0;
 	while (lst != NULL)
