+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: siferrar <siferrar@student.le-101.fr>      +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/10/09 09:16:21 by siferrar     #+#   ##    ##    #+#       */
-/*   Updated: 2019/10/19 00:45:32 by siferrar    ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/10/28 15:24:39 by siferrar    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -15,12 +15,12 @@
 
 void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	if ((dst == NULL && src == NULL))
 		return (NULL);
-	while (i < (int)n)
+	while (i < n)
 	{
 		((char *)dst)[i] = ((char *)src)[i];
 		i++;
